fix linter issues

+++ pkg/helpers/jobs_test.go
@@ -62,7 +62,7 @@ func TestWaitJobCompleted(t *testing.T) {
 		delay          time.Duration
 		expectError    bool
 		expectedResult bool
-		timeout        uint
+		timeout        int64
 	}
 
 	testCases := []TestCase{
